requirements.txt
python-dateutil
exif
openpyxl
pillow
pyyaml
